requirements.txt
Django==2.0.4
django-cors-headers==2.2.0
djangorestframework==3.8.2
pytz==2018.4
selenium==3.11.0
coverage==4.5.1